0.0.5: manage memory
Unfortunately TagLib stops working if we make enough objects without
freeing older ones. Add a free method to allow module users to free
unused objects. Might be a file handle limit.

We can probably remove this in the future if we keep the file closed
except for the initial read of the file, and then re-open it for
writing.